warn about pipeline attachments dropped by a shared pass
